# Conflicts:
#	zfs/cmd/Makefile.am

+++ zfs/cmd/cm/script_linux/cm_cnm_snapshot_backup.sh
@@ -7,6 +7,7 @@ CM_LOG_ENABLE=1
 CM_SSHD_CFG_FILE="/etc/ssh/sshd_config"
 CM_USER_ATTR_FILE="/etc/user_attr"
 CM_SSH_KEY_DIR="/etc/root/.ssh"
+CM_ETC_HOSTS="/etc/inet/hosts"
 
 function cm_cnm_snapshot_backup_init_sshd()
 {
@@ -14,7 +15,7 @@ function cm_cnm_snapshot_backup_init_sshd()
     cp ${CM_SSHD_CFG_FILE} "${CM_SSHD_CFG_FILE}.backup"
     rm $tmpfile 2>/dev/null
     cat ${CM_SSHD_CFG_FILE} | awk '{if($1=="PermitEmptyPasswords" || $1=="PermitRootLogin"){$2="yes";print;} else{print;};}' > $tmpfile
-	mv $tmpfile ${CM_SSHD_CFG_FILE}
+    mv $tmpfile ${CM_SSHD_CFG_FILE}
     return $CM_OK
 }
 
@@ -42,19 +43,40 @@ function cm_cnm_snapshot_backup_init_mtu()
 
 function cm_cnm_snapshot_backup_init_user_attr()
 {
-	local tmpfile=${CM_USER_ATTR_FILE}.tmp
-	rm $tmpfile 2>/dev/null
+    local tmpfile=${CM_USER_ATTR_FILE}.tmp
+    local bkupfile=${CM_USER_ATTR_FILE}.backup
+    
+    if [ ! -f $bkupfile ]; then
+        cp ${CM_USER_ATTR_FILE} ${bkupfile}
+    fi
+    rm $tmpfile 2>/dev/null
         cat ${CM_USER_ATTR_FILE} | awk 'BEGIN{FS=OFS=";"} {if($1~/root/){$6="type=normal";} print}' > $tmpfile
-	mv ${tmpfile} ${CM_USER_ATTR_FILE}
+    mv ${tmpfile} ${CM_USER_ATTR_FILE}
     return $CM_OK
 }
 
 function cm_cnm_snapshot_backup_term_user_attr()
 {
-	local tmpfile=${CM_USER_ATTR_FILE}.tmp
-	rm $tmpfile
-	cat ${CM_USER_ATTR_FILE} | awk 'BEGIN{FS=OFS=";"} {if($1~/root/){$6="type=role";} print}' > $tmpfile
-	mv ${tmpfile} ${CM_USER_ATTR_FILE}
+    local tmpfile=${CM_USER_ATTR_FILE}.tmp
+    
+    local bkupfile=${CM_USER_ATTR_FILE}.backup
+    
+    if [ -f ${bkupfile} ]; then
+        mv ${bkupfile} ${CM_USER_ATTR_FILE}
+        return $CM_OK
+    fi
+
+    rm $tmpfile
+
+    #illumos²»Ö§³Ö
+    local ostype=`cm_os_type_get`
+    if [ $ostype -eq ${CM_OS_TYPE_SOLARIS} ];then
+        cat ${CM_USER_ATTR_FILE} | awk 'BEGIN{FS=OFS=";"} {if($1~/root/){$6="type=role";} print}' > $tmpfile
+    elif [ $ostype -eq ${CM_OS_TYPE_ILLUMOS} ];then
+        cat ${CM_USER_ATTR_FILE} | awk 'BEGIN{FS=OFS=";"} {if($1~/root/){$6="";} print}' > $tmpfile
+    fi
+    
+    mv ${tmpfile} ${CM_USER_ATTR_FILE}
     return $CM_OK
 }
 
@@ -90,8 +112,8 @@ function cm_cnm_snapshot_backup_init_hosts()
 	local lname=$(hostname)
 	local rname=$(timeout 1 ceres_exec $2 'hostname')
 	[ -z "$rname" ]&&return $CM_FAIL
-	echo "$1 $lname loghost" >> /etc/hosts
-	echo "$2 $rname loghost" >> /etc/hosts
+	echo "$1 $lname loghost" >> ${CM_ETC_HOSTS}
+	echo "$2 $rname loghost" >> ${CM_ETC_HOSTS}
 	return $CM_OK
 }
 
@@ -100,8 +122,8 @@ function cm_cnm_snapshot_backup_term_hosts()
 	local local_ip=$1
 	local remote_ip=$2
 	local tmpfile="/etc/hosts.tmp"
-	sed "/$local_ip/d" /etc/hosts > $tmpfile
-	sed "/$remote_ip/d" $tmpfile > /etc/hosts
+	sed "/$local_ip/d" ${CM_ETC_HOSTS} > $tmpfile
+	sed "/$remote_ip/d" $tmpfile > ${CM_ETC_HOSTS}
 	mv $tmpfile
 	return $CM_OK
 }
